+++ Heimdal/lib/kadm5/ent_setup.c
@@ -169,12 +169,12 @@ _kadm5_setup_entry(kadm5_server_context *context,
     if(mask & KADM5_MAX_LIFE) {
 	if(princ_mask & KADM5_MAX_LIFE) {
 	    if(princ->max_life)
-	      set_value(ent->entry.max_life, princ->max_life);
+	      set_value(ent->entry.max_life, (unsigned int)princ->max_life);
 	    else
 	      set_null(ent->entry.max_life);
 	} else if(def_mask & KADM5_MAX_LIFE) {
 	    if(def->max_life)
-	      set_value(ent->entry.max_life, def->max_life);
+	      set_value(ent->entry.max_life, (unsigned int)def->max_life);
 	    else
 	      set_null(ent->entry.max_life);
 	}
@@ -191,12 +191,12 @@ _kadm5_setup_entry(kadm5_server_context *context,
     if(mask & KADM5_MAX_RLIFE) {
 	if(princ_mask & KADM5_MAX_RLIFE) {
 	  if(princ->max_renewable_life)
-	    set_value(ent->entry.max_renew, princ->max_renewable_life);
+	    set_value(ent->entry.max_renew, (unsigned int)princ->max_renewable_life);
 	  else
 	    set_null(ent->entry.max_renew);
 	} else if(def_mask & KADM5_MAX_RLIFE) {
 	  if(def->max_renewable_life)
-	    set_value(ent->entry.max_renew, def->max_renewable_life);
+	    set_value(ent->entry.max_renew, (unsigned int)def->max_renewable_life);
 	  else
 	    set_null(ent->entry.max_renew);
 	}
