+++ Heimdal/lib/krb5/prompter_posix.c
@@ -52,7 +52,7 @@ krb5_prompter_posix (krb5_context context,
     for (i = 0; i < num_prompts; ++i) {
 	if (prompts[i].hidden) {
 	    if(UI_UTIL_read_pw_string(prompts[i].reply->data,
-				  prompts[i].reply->length,
+				  (int)prompts[i].reply->length,
 				  prompts[i].prompt,
 				  0))
 	       return 1;
@@ -62,7 +62,7 @@ krb5_prompter_posix (krb5_context context,
 	    fputs (prompts[i].prompt, stdout);
 	    fflush (stdout);
 	    if(fgets(prompts[i].reply->data,
-		     prompts[i].reply->length,
+		     (int)prompts[i].reply->length,
 		     stdin) == NULL)
 		return 1;
 	    s[strcspn(s, "\n")] = '\0';
